requirements.txt
bert-extractive-summarizer==0.7.1
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
decorator==4.4.2
en-core-web-sm==3.0.0
filelock==3.0.12
flashtext==2.7
Flask==1.1.2
future==0.18.2
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
murmurhash==1.0.5
networkx==2.5.1
nltk==3.6.2
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pathy==0.5.2
pke==1.8.1
preshed==3.0.5
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywsd==1.2.4
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
sentencepiece==0.1.95
six==1.16.0
sklearn==0.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
thinc==8.0.3
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.5.1
typer==0.3.2
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
wn==0.0.23
zipp==3.4.1
